+++ cli/src/commit.rs
@@ -3,14 +3,11 @@ use atomic_lib::{errors::AtomicResult, Storelike};
 
 /// Apply a Commit using the Set method - create or update a value in a resource
 pub fn set(context: &Context) -> AtomicResult<()> {
-    let subcommand = "set";
-    let matches = context.matches.clone();
-    let subcommand_matches = matches.subcommand_matches(subcommand).clone().unwrap();
-    let subject = argument_to_url(context, subcommand, "subject")?;
-    let prop = argument_to_string(context, subcommand, "property")?;
-    let val = subcommand_matches.value_of("value").unwrap();
+    let subject = argument_to_url(context, "subject")?;
+    let prop = argument_to_string(context, "property")?;
+    let val = argument_to_string(context, "value")?;
     let mut commit_builder = atomic_lib::commit::CommitBuilder::new(subject);
-    commit_builder.set(prop, val.into());
+    commit_builder.set(prop, val);
     post(context, commit_builder)?;
     Ok(())
 }
@@ -18,43 +15,36 @@ pub fn set(context: &Context) -> AtomicResult<()> {
 /// Apply a Commit using the Set method, where the value is edited in the user's text editor.
 #[cfg(feature = "native")]
 pub fn edit(context: &Context) -> AtomicResult<()> {
-    let subcommand = "edit";
-    let subject = argument_to_url(context, subcommand, "subject")?;
-    let prop = argument_to_string(context, subcommand, "property")?;
-    let current_val = context
-        .store
-        .get_resource(&subject)?
-        .get_shortname(&prop, &context.store)?
-        .to_string();
+    let subject = argument_to_url(context, "subject")?;
+    let prop = argument_to_string(context,  "property")?;
+    let mut resource = context.store.get_resource(&subject)?;
+    let current_val = resource.get_shortname(&prop, &context.store)?.to_string();
     let edited = edit::edit(current_val)?;
-    let mut commit_builder = atomic_lib::commit::CommitBuilder::new(subject);
-    commit_builder.set(prop, edited);
-    post(context, commit_builder)?;
+    resource.set_propval_shortname(&prop, &edited, &context.store)?;
+    post(context, resource.get_commit_builder().clone())?;
     Ok(())
 }
 
 /// Apply a Commit using the Remove method - removes a property from a resource
 pub fn remove(context: &Context) -> AtomicResult<()> {
-    let subcommand = "remove";
-    let subject = argument_to_url(context, subcommand, "subject")?;
-    let prop = argument_to_string(context, subcommand, "property")?;
-    let mut commit_builder = atomic_lib::commit::CommitBuilder::new(subject);
-    commit_builder.remove(prop);
-    post(context, commit_builder)?;
+    let subject = argument_to_url(context, "subject")?;
+    let prop = argument_to_string(context, "property")?;
+    let mut resource = context.store.get_resource(&subject)?;
+    resource.remove_propval_shortname(&prop, &context.store)?;
+    post(context, resource.get_commit_builder().clone())?;
     Ok(())
 }
 
 /// Apply a Commit using the destroy method - removes a resource
 pub fn destroy(context: &Context) -> AtomicResult<()> {
-    let subcommand = "destroy";
-    let subject = argument_to_url(context, subcommand, "subject")?;
+    let subject = argument_to_url(context, "subject")?;
     let mut commit_builder = atomic_lib::commit::CommitBuilder::new(subject);
     commit_builder.destroy(true);
     post(context, commit_builder)?;
     Ok(())
 }
 
-/// Posts the Commit and applies it to the server
+/// Signs the Commit, Posts it and applies it to the server
 fn post(context: &Context, commit_builder: atomic_lib::commit::CommitBuilder) -> AtomicResult<()> {
     context.get_write_context();
     let agent = context
@@ -66,8 +56,10 @@ fn post(context: &Context, commit_builder: atomic_lib::commit::CommitBuilder) ->
     Ok(())
 }
 
-fn argument_to_string(context: &Context, subcommand: &str, argument: &str) -> AtomicResult<String> {
-    let subcommand_matches = context.matches.subcommand_matches(subcommand).unwrap();
+/// Parses a single argument as string
+fn argument_to_string(context: &Context, argument: &str) -> AtomicResult<String> {
+    let command_name = context.matches.subcommand_name().unwrap();
+    let subcommand_matches = context.matches.subcommand_matches(command_name).unwrap();
     let user_arg = subcommand_matches
         .value_of(argument)
         .ok_or(format!("No argument value for {} found", argument))?;
@@ -75,8 +67,9 @@ fn argument_to_string(context: &Context, subcommand: &str, argument: &str) -> At
 }
 
 /// Parses a single argument (URL or Bookmark), should return a valid URL
-fn argument_to_url(context: &Context, subcommand: &str, argument: &str) -> AtomicResult<String> {
-    let subcommand_matches = context.matches.subcommand_matches(subcommand).unwrap();
+fn argument_to_url(context: &Context, argument: &str) -> AtomicResult<String> {
+    let command_name = context.matches.subcommand_name().unwrap();
+    let subcommand_matches = context.matches.subcommand_matches(command_name).unwrap();
     let user_arg = subcommand_matches
         .value_of(argument)
         .ok_or(format!("No argument value for {} found", argument))?;