+++ lib/examples/basic.rs
@@ -32,7 +32,7 @@ fn main() {
     // Let's make a new Property instance! Let's create "age".
     let mut new_property = atomic_lib::Resource::new_instance("https://atomicdata.dev/classes/Property", &store).unwrap();
     // And add a description for that Property
-    new_property.set_propval_by_shortname("description", "the age of a person", &store).unwrap();
+    new_property.set_propval_shortname("description", "the age of a person", &store).unwrap();
     // A subject URL for the new resource has been created automatically.
     let subject = new_property.get_subject().clone();
     // Now we need to make sure these changes are also applied to the store.
@@ -45,8 +45,8 @@ fn main() {
     // But.. when we commit, we get an error!
     // Because we haven't set all the properties required for the Property class.
     // We still need to set `shortname` and `datatype`.
-    new_property.set_propval_by_shortname("shortname", "age", &store).unwrap();
-    new_property.set_propval_by_shortname("datatype", atomic_lib::urls::INTEGER, &store).unwrap();
+    new_property.set_propval_shortname("shortname", "age", &store).unwrap();
+    new_property.set_propval_shortname("datatype", atomic_lib::urls::INTEGER, &store).unwrap();
     new_property.save(&store).unwrap();
     // Now the changes to the resource applied to the store, and we can fetch the newly created resource!
     let fetched_new_resource = store.get_resource(&subject).unwrap();